Refactor navigation in app
It's now completely internal to the
app module.
