--- base/src/main/java/app/tivi/Screen.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2021 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package app.tivi
-
-enum class Screen(val route: String) {
-    Discover("discover"),
-    Following("following"),
-    Trending("trending"),
-    Popular("popular"),
-    ShowDetails("show/{showId}"),
-    EpisodeDetails("episode/{episodeId}"),
-    RecommendedShows("recommendedshows"),
-    Watched("watched"),
-    Search("search"),
-    Account("account"),
-}
